requirements.txt
soundfile
librosa
Werkzeug
Flask
google-cloud-speech